fix: OCR-feature CCITT diagnostic tests use fax 0.2's u16 decode_g4 API
fax was bumped to 0.3 (Cargo.toml already pins fax = "0.3"), which
changed decode_g4's width/height parameters from u16/Option<u16> to
u32/Option<u32>. Production code (ccitt_bilevel.rs) was updated for
this at the time, but two ocr-feature-gated diagnostic test files
(test_ccitt_edge_cases.rs, test_ccitt_decoder_diagnosis.rs) still
truncated to u16 before calling decode_g4, which no longer compiles
against fax 0.3's signature.

Neither `cargo test` (default features) nor the rendering/fips,icc
combos compile these files, since they're gated behind the `ocr`
feature — hence why this went unnoticed until built with `--features
rendering,barcodes,signatures,ocr`.

Drop the stale u16 truncation and pass the already-u32 width/height
straight through, matching what ccitt_bilevel.rs already does.

+++ tests/test_ccitt_edge_cases.rs
@@ -61,8 +61,8 @@ mod ccitt_edge_cases {
                             println!("\n  Testing fax decoder with leading zeros stripped:");
                             test_fax_decoder_on_data(
                                 &pixels[leading_zeros..],
-                                image.width() as u16,
-                                Some(image.height() as u16),
+                                image.width(),
+                                Some(image.height()),
                                 &format!("Image {} (zeros stripped)", idx),
                             );
                         } else {
@@ -77,7 +77,7 @@ mod ccitt_edge_cases {
         }
     }
 
-    fn test_fax_decoder_on_data(data: &[u8], width: u16, height: Option<u16>, _label: &str) {
+    fn test_fax_decoder_on_data(data: &[u8], width: u32, height: Option<u32>, _label: &str) {
         use fax::decoder;
 
         let mut output_count = 0;
